memcg: do not destroy kmem caches on css offline
Currently, each kmem active memory cgroup has its own set of kmem caches. 
The caches are only used by the memory cgroup they were created for, so
when the cgroup is taken offline they must be destroyed.  However, we
can't easily destroy all the caches on css offline, because they still may
contain objects accounted to the cgroup.  Actually, we don't bother
destroying busy caches on css offline at all, effectively leaking them. 
To make this scheme work as it was intended to, we have to introduce a
kind of asynchronous caches destruction, which is going to be quite a
complex stuff, because we'd have to handle a lot of various race
conditions.  And even if we manage to solve them all, kmem caches created
for memory cgroups that are now dead will be dangling indefinitely long
wasting memory.

In this patch set I implement a different approach, which can be
described by the following statements:

 1. Never destroy per memcg kmem caches (except the root cache is
    destroyed, of course).

 2. Reuse kmemcg_id and therefore the set of per memcg kmem caches left
    from a dead memory cgroup.

 3. After allocating a kmem object, check if the slab is accounted to
    the proper (i.e. current) memory cgroup. If it doesn't recharge it.

The benefits are:

 - It's much simpler than what we have now, even though the current
   implementation is incomplete.

 - The number of per cgroup caches of the same kind cannot be be greater
   than the maximal number of online kmem active memory cgroups that
   have ever existed simultaneously. Currently it is unlimited, which is
   really bad.

 - Once a new memory cgroup starts using a cache that was used by a dead
   cgroup before, it will be recharging slabs accounted to the dead
   cgroup while allocating objects from the cache. Therefore all
   references to the old cgroup will be put sooner or later, and it will
   be freed. Currently, cgroups that have kmem objects accounted to them
   on css offline leak for good.



This patch (of 8):

Currently, we try to destroy per memcg kmem caches on css offline.  Since
a cache can contain active objects when the memory cgroup is removed, we
can't destroy all caches immediately and therefore should introduce
asynchronous destruction for this scheme to work properly.  However, this
requires a lot of trickery and complex synchronization stuff, so I'm
planning to go another way.  I'm going to reuse caches left from dead
memory cgroups instead of recreating them.  This patch makes the first
step in this direction: it removes caches destruction from css offline.

Signed-off-by: Vladimir Davydov <vdavydov@parallels.com>
Cc: Johannes Weiner <hannes@cmpxchg.org>
Cc: Michal Hocko <mhocko@suse.cz>
Cc: Christoph Lameter <cl@linux.com>
Cc: Pekka Enberg <penberg@kernel.org>
Cc: David Rientjes <rientjes@google.com>
Cc: Joonsoo Kim <iamjoonsoo.kim@lge.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>