slab: recharge slab pages to the allocating memory cgroup
Since we now reuse per cgroup kmem caches, the slab we allocate an object
from may be accounted to a dead memory cgroup.  If we leave such a slab
accounted to a dead cgroup, we risk pinning the cgroup forever, so we
introduce a new function, memcg_kmem_recharge_slab, which is to be called
in the end of kmalloc.  It recharges the new object's slab to the current
cgroup unless it is already charged to it.

Signed-off-by: Vladimir Davydov <vdavydov@parallels.com>
Cc: Johannes Weiner <hannes@cmpxchg.org>
Cc: Michal Hocko <mhocko@suse.cz>
Cc: Christoph Lameter <cl@linux.com>
Cc: Pekka Enberg <penberg@kernel.org>
Cc: David Rientjes <rientjes@google.com>
Cc: Joonsoo Kim <iamjoonsoo.kim@lge.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>