param: fix crash on bad kernel arguments
Currently if the user passes an invalid value on the kernel command line
then the kernel will crash during argument parsing.  On most systems this
is very hard to debug because the console hasn't been initialized yet.

This is a regression due to commit 51e158c12aca ("param: hand arguments
after -- straight to init") which, in response to the systemd debug
controversy, made it possible to explicitly pass arguments to init.  To
achieve this parse_args() was extended from simply returning an error code
to returning a pointer.  Regretably the new init args logic does not
perform a proper validity check on the pointer resulting in a crash.

This patch fixes the validity check.  Should the check fail then no
arguments will be passed to init.  This is reasonable and matches how the
kernel treats its own arguments (i.e.  no error recovery).

Signed-off-by: Daniel Thompson <daniel.thompson@linaro.org>
Cc: Rusty Russell <rusty@rustcorp.com.au>
Cc: <stable@vger.kernel.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>