rtc: ds1307: add device tree support
Note that alarm interrupt and trickle charger are not handled (I have no
hardware with these features).

Only the "common" compatible string for that driver (maxim,ds1307) is
documented, so checkpatch complains for the rest.  If it is better to
document every compatible, I will post a new series with documentation
files pointing to the maxim,ds1307.txt.

Signed-off-by: Tomas Novotny <tomas@novotny.cz>
Cc: Alessandro Zummo <a.zummo@towertech.it>
Cc: Grant Likely <grant.likely@linaro.org>
Cc: Rob Herring <robh+dt@kernel.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>