drivers/char/mem.c: make /dev/mem an optional device
Make the /dev/mem device optional using the CONFIG_DEVMEM option.

This brings /dev/mem in line with /dev/kmem so it can de disabled. 
Disabling /dev/mem is preferable on devices that do not require direct
access to memory in order to improve security.  This is especially true on
platforms where no restiction is placed on the areas of memory that can be
accessed(for example mips) without custom modifcations to the kernel
source.

The implementation of this functionality declares the mem_fops structure
as __maybe_unused to eliminate the need for #ifdefs throughout the file.

Follow up changes are present to change CONFIG_DEVKMEM and CONFIG_DEVPORT
to use __maybe_unused.  This simplifies the code and make the
implementations consistent.


This patch (of 3):

Adds Kconfig option CONFIG_DEVMEM that allows the /dev/mem device to be
disabled.

Option defaults to /dev/mem enabled.

Signed-off-by: Rob Ward <robert.ward114@googlemail.com>
Acked-by: Arnd Bergmann <arnd@arndb.de>
Cc: Greg KH <greg@kroah.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>